+++ flows/podman/podman_flows.py
@@ -7,6 +7,7 @@ from prefect.utilities.processutils import run_process
 
 from flows.logger import setup_logger
 from flows.podman.schema import PodmanParams
+from flows.credentials import add_credentials_to_io_parameters
 
 
 @flow(name="Podman flow")
@@ -27,6 +28,9 @@ async def launch_podman(
 
     # Append current flow run id
     podman_params.params["io_parameters"]["uid_save"] = current_flow_run_id
+
+    # Add credentials to io_parameters at the child flow level
+    podman_params.params = add_credentials_to_io_parameters(podman_params.params)
 
     # Create temporary file for parameters
     with tempfile.NamedTemporaryFile(mode="w+t") as temp_file:
