+++ flows/slurm/slurm_flows.py
@@ -8,6 +8,7 @@ from prefect.utilities.processutils import run_process
 
 from flows.logger import setup_logger
 from flows.slurm.schema import SlurmParams
+from flows.credentials import add_credentials_to_io_parameters
 
 
 @flow(name="launch_slurm")
@@ -25,6 +26,9 @@ async def launch_slurm(
 
     # Append current flow run id
     slurm_params.params["io_parameters"]["uid_save"] = current_flow_run_id
+
+    # Add credentials to io_parameters at the child flow level
+    slurm_params.params = add_credentials_to_io_parameters(slurm_params.params)
 
     # Create temporary file for parameters
     with tempfile.NamedTemporaryFile(mode="w+t", dir=".") as temp_file:
